+++ src/Modules/Admin/Widgets/Forms/AccountActiveForm.php
@@ -35,10 +35,7 @@ class AccountActiveForm extends ActiveForm
                 $this->getTimezoneField(),
             ],
             [
-                $this->getFirstNameField(),
-                $this->getLastNameField(),
-                $this->getCityField(),
-                $this->getCountryField(),
+                ...$this->getUserCustomAttributeFields(),
             ],
         ];